+++ test/test_base.py
@@ -1,27 +1,28 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=W0312
-'''Test pyonep library.'''
+"""Test pyonep library."""
 from __future__ import unicode_literals
-import json
-import md5
-import os
 from unittest import TestCase
 
-# import vcr
 import requests
 
 from pyonep import onep
 
+
 CIK_FNTN_URL = 'https://cik.herokuapp.com/api'
-print(os.path.abspath(__file__))
-CIK_FNTN_VENDOR_ID = md5.md5(os.path.abspath(__file__)).hexdigest()
-print("VENDOR_ID: {0}".format(CIK_FNTN_VENDOR_ID))
-# sys.exit()
-CIK_FNTN_CASS = 'vcr_cassettes/cik_fntn_cassette.yaml'
-# myvcr = vcr.VCR(
-#     match_on=['method', 'scheme', 'uri', 'query', 'body'],
-#     record_mode='new_episodes'
-# )
+
+def create_test_client():
+    return TestBase.createTestClient()
+
+
+class TestClient():
+    def __init__(self, config):
+        self.config = config
+        self.cik = config['cik']
+        self.rid = config['rid']
+        self.vendor = config['vendor']
+        self.vendortoken = config['vendortoken']
+
 
 class TestBase(TestCase):
     """
@@ -32,20 +33,25 @@ class TestBase(TestCase):
     __test__ = False
 
     @classmethod
+    def createTestClient(cls):
+        r = requests.get(CIK_FNTN_URL+'/create?vendor')
+        r.raise_for_status()
+        config = r.json()
+        return TestClient(config)
+
+    @classmethod
     def setUpClass(cls):
         cls.session = requests.Session()
-        # with myvcr.use_cassette(CIK_FNTN_CASS):
-        r = requests.get(CIK_FNTN_URL+'/create?vendor') #='+CIK_FNTN_VENDOR_ID)
-        print(r.status_code, r.text)
+        r = requests.get(CIK_FNTN_URL+'/create?vendor')
+        r.raise_for_status()
         cls.config = r.json()
-        print("Here's the config from CIK Fountain:\n{0}".format(json.dumps(cls.config, indent=2)))
 
     @classmethod
     def tearDownClass(cls):
         pass
 
     def setUp(self):
-        '''Create a device in the portal to test'''
+        """Create a device in the portal to test"""
         self.portalcik = self.config['cik']
         self.vendorname = self.config['vendor']
         self.vendortoken = self.config['vendortoken']
@@ -85,7 +91,7 @@ class TestBase(TestCase):
         self.cik = response['key']
 
     def tearDown(self):
-        '''Clean up any test client'''
+        """Clean up any test client"""
         self.onep.drop(self.portalcik, self.rid)
 
     def makeClient(self, cik):