+++ test/test_rpc.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=W0312
-'''Test pyonep RPC'''
+"""Test pyonep RPC"""
 from __future__ import unicode_literals
+import doctest
+import re
 
-import test_base
+from test import test_base
 
 
 class TestRPC(test_base.TestBase):
@@ -13,9 +15,13 @@ class TestRPC(test_base.TestBase):
     # yes, test this
     __test__ = True
 
+    def test_doc_examples(self):
+        """Test documentation examples"""
+        doctest.testfile('../docs/examples.md')
+
     # @myvcr.use_cassette('vcr_cassettes/test_move.yaml')
     def test_move(self):
-        '''Test move command'''
+        """Test move command"""
         #    self.cik
         #       |
         #      cik2
@@ -33,3 +39,73 @@ class TestRPC(test_base.TestBase):
         #     |    |
         #    cik3 cik2
         self.assertTrue(ok, 'move succeeded')
+
+    def test_create(self):
+        """Test create API call"""
+        isok, response = self.onep.create(
+            self.cik,
+            'client',
+            {
+                'writeinterval': 'inherit',
+                'name': 'testclient',
+                'visibility': 'parent',
+                'limits': {
+                    'dataport': 'inherit',
+                    'datarule': 'inherit',
+                    'dispatch': 'inherit',
+                    'disk': 'inherit',
+                    'io': 'inherit',
+                    'share': 'inherit',
+                    'client': 'inherit',
+                    'sms': 'inherit',
+                    'sms_bucket': 'inherit',
+                    'email': 'inherit',
+                    'email_bucket': 'inherit',
+                    'http': 'inherit',
+                    'http_bucket': 'inherit',
+                    'xmpp': 'inherit',
+                    'xmpp_bucket': 'inherit'}
+            })
+        client_rid = response
+        self.assertTrue(isok, 'client creation succeeded')
+        self.assertTrue(re.match("^[0-9a-f]{40}$", client_rid), 'rid is formatted correctly')
+
+        isok, response = self.onep.info(
+            self.cik,
+            client_rid,
+            {'key': True}
+        )
+        client_cik = response['key']
+
+        # Add a dataport
+        isok, response = self.onep.create(
+            client_cik,
+            'dataport',
+            {
+                'format': 'string',
+                'retention': {
+                    'count': 'infinity',
+                    'duration': 'infinity',
+                },
+                'limits': {
+                    'dataport': 'inherit',
+                    'datarule': 'inherit',
+                    'dispatch': 'inherit',
+                    'disk': 'inherit',
+                    'io': 'inherit',
+                    'share': 'inherit',
+                    'client': 'inherit',
+                    'sms': 'inherit',
+                    'sms_bucket': 'inherit',
+                    'email': 'inherit',
+                    'email_bucket': 'inherit',
+                    'http': 'inherit',
+                    'http_bucket': 'inherit',
+                    'xmpp': 'inherit',
+                    'xmpp_bucket': 'inherit',
+                }
+            }
+        )
+        dataport_rid = response
+        self.assertTrue(isok, 'dataport creation succeeded')
+        self.assertTrue(re.match("^[0-9a-f]{40}$", dataport_rid), 'rid is formatted correctly')